+++ src/geometryXVx/rhs/collisions_intra.cpp
@@ -264,9 +264,9 @@ DFieldSpXVx CollisionsIntra::operator()(DFieldSpXVx allfdistribu, double dt) con
 
     IdxRangeSpX grid_sp_x(get_idx_range<Species, GridX>(allfdistribu));
     // density and temperature
-    DFieldMemSpX density_alloc(grid_sp_x);
-    DFieldMemSpX fluid_velocity_alloc(grid_sp_x);
-    DFieldMemSpX temperature_alloc(grid_sp_x);
+    DFieldMemSpX density_alloc("density (collisions_intra.cpp)", grid_sp_x);
+    DFieldMemSpX fluid_velocity_alloc("fluid_velocity (collisions_intra.cpp)", grid_sp_x);
+    DFieldMemSpX temperature_alloc("temperature (collisions_intra.cpp)", grid_sp_x);
     DFieldSpX density = get_field(density_alloc);
     DFieldSpX fluid_velocity = get_field(fluid_velocity_alloc);
     DFieldSpX temperature = get_field(temperature_alloc);
@@ -299,7 +299,7 @@ DFieldSpXVx CollisionsIntra::operator()(DFieldSpXVx allfdistribu, double dt) con
             });
 
     // collision frequency
-    DFieldMemSpX collfreq_alloc(grid_sp_x);
+    DFieldMemSpX collfreq_alloc("collfreq (collisions_intra.cpp)", grid_sp_x);
     DFieldSpX collfreq = get_field(collfreq_alloc);
     compute_collfreq(
             collfreq,
@@ -308,7 +308,7 @@ DFieldSpXVx CollisionsIntra::operator()(DFieldSpXVx allfdistribu, double dt) con
             get_const_field(temperature));
 
     // diffusion coefficient
-    DFieldMem<IdxRangeSpXVx_ghosted> Dcoll_alloc(m_mesh_ghosted);
+    DFieldMem<IdxRangeSpXVx_ghosted> Dcoll_alloc("Dcoll (collisions_intra.cpp)", m_mesh_ghosted);
     DField<IdxRangeSpXVx_ghosted> Dcoll = get_field(Dcoll_alloc);
     compute_Dcoll<GhostedVx>(
             Dcoll,
@@ -316,7 +316,8 @@ DFieldSpXVx CollisionsIntra::operator()(DFieldSpXVx allfdistribu, double dt) con
             get_const_field(density),
             get_const_field(temperature));
 
-    DFieldMem<IdxRangeSpXVx_ghosted> dvDcoll_alloc(m_mesh_ghosted);
+    DFieldMem<IdxRangeSpXVx_ghosted>
+            dvDcoll_alloc("dvDcoll (collisions_intra.cpp)", m_mesh_ghosted);
     DField<IdxRangeSpXVx_ghosted> dvDcoll = get_field(dvDcoll_alloc);
     compute_dvDcoll<GhostedVx>(
             dvDcoll,
@@ -324,7 +325,9 @@ DFieldSpXVx CollisionsIntra::operator()(DFieldSpXVx allfdistribu, double dt) con
             get_const_field(density),
             get_const_field(temperature));
 
-    DFieldMem<IdxRangeSpXVx_ghosted_staggered> Dcoll_staggered_alloc(m_mesh_ghosted_staggered);
+    DFieldMem<IdxRangeSpXVx_ghosted_staggered> Dcoll_staggered_alloc(
+            "Dcoll_staggered (collisions_intra.cpp)",
+            m_mesh_ghosted_staggered);
     DField<IdxRangeSpXVx_ghosted_staggered> Dcoll_staggered = get_field(Dcoll_staggered_alloc);
     compute_Dcoll<GhostedVxStaggered>(
             Dcoll_staggered,
@@ -333,21 +336,21 @@ DFieldSpXVx CollisionsIntra::operator()(DFieldSpXVx allfdistribu, double dt) con
             get_const_field(temperature));
 
     // kernel maxwellian fluid moments
-    DFieldMemSpX Vcoll_alloc(grid_sp_x);
-    DFieldMemSpX Tcoll_alloc(grid_sp_x);
+    DFieldMemSpX Vcoll_alloc("Vcoll (collisions_intra.cpp)", grid_sp_x);
+    DFieldMemSpX Tcoll_alloc("Tcoll (collisions_intra.cpp)", grid_sp_x);
     DFieldSpX Vcoll = get_field(Vcoll_alloc);
     DFieldSpX Tcoll = get_field(Tcoll_alloc);
     compute_Vcoll_Tcoll<GhostedVx>(Vcoll, Tcoll, get_const_field(allfdistribu), Dcoll, dvDcoll);
 
     // convection coefficient Nucoll
-    DFieldMem<IdxRangeSpXVx_ghosted> Nucoll_alloc(m_mesh_ghosted);
+    DFieldMem<IdxRangeSpXVx_ghosted> Nucoll_alloc("Nucoll (collisions_intra.cpp)", m_mesh_ghosted);
     DField<IdxRangeSpXVx_ghosted> Nucoll = get_field(Nucoll_alloc);
     compute_Nucoll<GhostedVx>(Nucoll, Dcoll, get_const_field(Vcoll), get_const_field(Tcoll));
 
     // matrix coefficients
-    DFieldMemSpXVx AA_alloc(get_idx_range(allfdistribu));
-    DFieldMemSpXVx BB_alloc(get_idx_range(allfdistribu));
-    DFieldMemSpXVx CC_alloc(get_idx_range(allfdistribu));
+    DFieldMemSpXVx AA_alloc("AA (collisions_intra.cpp)", get_idx_range(allfdistribu));
+    DFieldMemSpXVx BB_alloc("BB (collisions_intra.cpp)", get_idx_range(allfdistribu));
+    DFieldMemSpXVx CC_alloc("CC (collisions_intra.cpp)", get_idx_range(allfdistribu));
     DFieldSpXVx AA = get_field(AA_alloc);
     DFieldSpXVx BB = get_field(BB_alloc);
     DFieldSpXVx CC = get_field(CC_alloc);
@@ -355,7 +358,7 @@ DFieldSpXVx CollisionsIntra::operator()(DFieldSpXVx allfdistribu, double dt) con
 
 
     // rhs vector coefficient
-    DFieldMemSpXVx RR_alloc(get_idx_range(allfdistribu));
+    DFieldMemSpXVx RR_alloc("RR (collisions_intra.cpp)", get_idx_range(allfdistribu));
     DFieldSpXVx RR = get_field(RR_alloc);
     compute_rhs_vector(
             RR,