+++ CHANGELOG.md
@@ -7,6 +7,46 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [v0.7.0] - 2026-03-18
+
+### Added
+
+- Add `SplineInterpolator` class to group a compatible DDC spline builder and evaluator.
+- Add `LagrangeInterpolator` class to group a compatible builder and evaluator.
+- Add constructors taking a `concepts::Interpolator` class instance for:
+  - `BslAdvectionSpatial`
+  - `BslAdvectionVelocity`
+  - `BslAdvection1D`
+  - `FEM1DPoissonSolver`
+- Add a constructor to `BslAdvection1D` to use the same interpolator type for both the advected function and the advection field.
+- Add new constructors for `VectorFieldMem` to support labels as used in DDC and Kokkos.
+- Added index range assertions in `Quadrature`.
+
+### Fixed
+
+- Add missing `enable_tensor_type` for `CartesianLeviCivitaTensor` and LeviCivita `size()`.
+
+### Changed
+
+- Add Koliop in all toolchains.
+- Changed spline boundary condition in velocity dimensions in the XVx and XYVxVy geometries to use homogeneous Hermite boundary conditions.
+- Change `BslAdvectionVelocity` to stop providing values for the derivatives at the boundaries.
+- Add labels to all parallel constructs and many variable allocations.
+- Completed the porting of `PolarSplineFEMPoissonLikeAssembler` to GPU (less memory, fewer data transfers, faster execution).
+
+### Deprecated
+
+- Deprecate constructors taking both builders and evaluators n favour of constructors taking Interpolator classes for:
+  - `BslAdvectionSpatial`
+  - `BslAdvectionVelocity`
+  - `BslAdvection1D`
+  - `FEM1DPoissonSolver`
+
+### Removed
+
+- Remove deprecated `Lagrange` class.
+- Remove Koliop submodule.
+
 ## [v0.6.2] - 2026-03-06
 
 ### Fixed